requirements.txt
aioredis==1.3.1
amqp==5.1.1
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
autopep8==1.6.0
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.5.18.1
#cffi==1.15.0
channels==3.0.4
channels-redis==3.4.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.5
django-anymail==8.6
django-cors-headers==3.12.0
django-filter==21.1
django-sendgrid-v5==1.2.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
ecdsa==0.14.1
google-auth==2.6.6
gunicorn==20.1.0
h11==0.12.0
hiredis==2.1.0
httpcore==0.15.0
httpx==0.23.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
kubernetes==23.6.0
Markdown==3.3.7
MarkupSafe==2.1.1
msgpack==1.0.4
oauthlib==3.2.0
octokitpy==0.15.0
octokitpy-routes==2.5.0
packaging==21.3
prompt-toolkit==3.0.31
#psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-http-client==3.3.7
python-jose==3.2.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
rave-python==1.2.16
redis==4.3.4
requests==2.25.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.8
sendgrid==6.9.7
service-identity==21.1.0
simplejson==3.18.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
toml==0.10.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.3.2
wrapt==1.14.1
zope.interface==5.4.0
